Update: move update_feed_path_configuration
