+++ src/wazuh_testing/modules/vulnerability_detector/packages.py
@@ -83,45 +83,27 @@ SUSE_SYSTEM_PACKAGE = {
             "format": "rpm",
             "vendor": "SUSE LLC <https://www.suse.com/>"
         }
+    ],
+    'SLES11': [
+        {
+            "name": "sle-module-basesystem-release",
+            "version": "11.2",
+            "format": "rpm",
+            "vendor": "SUSE LLC <https://www.suse.com/>"
+        }
+    ],
+    'SLES12': [
+        {
+            "name": "sle-module-basesystem-release",
+            "version": "12.2",
+            "format": "rpm",
+            "vendor": "SUSE LLC <https://www.suse.com/>"
+        }
     ]
 }
 
 
 # Packages and Feed related functions
-def update_feed_path_configurations(configurations: list, metadata: list, feeds_path: str) -> list:
-    """Replace feed path tags in the configuration template, using the metadata information.
-
-    Args:
-        configurations (list(dict)): List of configuration templates.
-        metadata (list(dict)): List of configuration templates metadata.
-        feeds_path (str): Absolute path where the feeds are located.
-
-    Returns:
-        list(dict): List of configurations with the feeds path updated.
-    """
-    new_configurations = deepcopy(configurations)
-
-    for index, _ in enumerate(configurations):
-        if 'json_feed' in metadata[index] and metadata[index]['json_feed'] is not None:
-            new_configurations[index] = json.loads(json.dumps(new_configurations[index]).
-                                                   replace(metadata[index]['json_feed_tag'],
-                                                   os.path.join(feeds_path, metadata[index]['provider_name'],
-                                                                metadata[index]['json_feed'])))
-
-        if 'oval_feed' in metadata[index] and metadata[index]['oval_feed'] is not None:
-            new_configurations[index] = json.loads(json.dumps(new_configurations[index]).
-                                                   replace(metadata[index]['oval_feed_tag'],
-                                                   os.path.join(feeds_path, metadata[index]['provider_name'],
-                                                                metadata[index]['oval_feed'])))
-
-        if 'nvd_feed_tag' in metadata[index] and 'nvd_feed' in metadata[index]:
-            new_configurations[index] = json.loads(json.dumps(new_configurations[index]).
-                                                   replace(metadata[index]['nvd_feed_tag'],
-                                                   os.path.join(feeds_path, 'nvd', metadata[index]['nvd_feed'])))
-
-    return new_configurations
-
-
 def insert_vulnerable_packages(agent_id: str = '000', vendor: str = 'Red Hat, Inc.', count: int = 5) -> None:
     """Insert vulnerable packages to an agent.
